(feature:Beneficiaries) updated some files as guided
